### 2022-10-07 #557 @trangdata
Presubmission inquiry - openalexR: interacts with OpenAlex API
Submitting Author Name: Trang Le
Submitting Author Github Handle: @trangdata
Repository:  https://github.com/massimoaria/openalexR
Submission type: Pre-submission

---

```
Package: openalexR
Title: Getting Bibliographic Records from 'OpenAlex' Database Using 'DSL'
    API
Version: 1.0.1
    person(given = "Massimo",
           family = "Aria",
           email = "aria@unina.it",
    person(given = "Corrado",
           family = "Cuccurullo",
           role = c("ctb"),
           email = "cuccurullocorrado@gmail.com",
    person(given = "Trang",
           family = "Le",
           role = "aut",
           email = "grixor@gmail.com",
    )
Description: A set of tools to extract bibliographic content from
    'OpenAlex' database using API <https://docs.openalex.org/api/>.
    https://massimoaria.github.io/openalexR/
    curl,
    httr,
    progress,
    tibble
    covr
    knitr
```

The package interacts with the [OpenAlex API](https://docs.openalex.org/api). Similar packages in the same category are rcrossref and rotl.

N/A

Anyone who wants to work in R to interact with the OpenAlex API to acquire information on publications, authors, etc., including researchers in the field of bibliometrics, text mining, etc. We including some nice example analyses in our README.

There are a couple of packages that we're aware of: https://github.com/KTH-Library/openalex and https://github.com/ekmaloney/openalexR. We think our package strikes the balance of complexity/flexibility, follow best API package practices, and offer additional useful functionality such as [snowball](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4211030/#:~:text=Background-,Snowballing,-involves%20recursively%20pursuing) search.

Yes.

None.
