### 2022-10-07 6d55810
Add is_paratext & is_retracted as columns in work2df
closes https://github.com/massimoaria/openalexR/issues/27
