+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   A_Brief_Introduction_to_openalexR: A_Brief_Introduction_to_openalexR.html
   About-the-tibble-output: About-the-tibble-output.html
   institution: institution.html
-last_built: 2022-10-07T13:51Z
+last_built: 2022-10-07T19:20Z
 urls:
   reference: https://massimoaria.github.io/openalexR/reference
   article: https://massimoaria.github.io/openalexR/articles
