+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 export(oa2bibliometrix)
 export(oa2df)
+export(oa_entities)
 export(oa_fetch)
 export(oa_query)
 export(oa_random)
