### 2022-10-10 e1e8073
minor updates to manual
