+++ man/authors2df.Rd
@@ -9,8 +9,8 @@ authors2df(data, verbose = TRUE)
 \arguments{
 \item{data}{List. Output of \code{oa_request}.}
 
-\item{verbose}{Logical. If TRUE, information about the querying process will
-be plotted on screen. Default to \code{verbose = FALSE}.}
+\item{verbose}{Logical. If TRUE, print information on querying process.
+Default to \code{verbose = FALSE}.}
 }
 \value{
 a data.frame.
