+++ man/snowball2df.Rd
@@ -1,19 +1,21 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/oa_snowball.R
-\name{to_disk}
-\alias{to_disk}
+% Please edit documentation in R/oa2df.R
+\name{snowball2df}
+\alias{snowball2df}
 \title{Flatten snowball result}
 \usage{
-to_disk(snowball)
+snowball2df(data, verbose = FALSE)
 }
 \arguments{
-\item{snowball}{List result from `oa_snowball`.}
+\item{data}{List result from `oa_snowball`.}
+
+\item{verbose}{Logical. If TRUE, print information on wrangling process.}
 }
 \value{
 Tibble/data.frame of works with additional columns:
 append `citing`, `backward_count`, `cited_by`, `forward_count`, `connection`,
 and `connection_count.` For each work/row, these counts are WITHIN one
-snowball search, and so `forward_count` <= `cited_by_count`.
+data search, and so `forward_count` <= `cited_by_count`.
 
 Consider the universe of all works linked to a set of starting works, (`oa_input = TRUE`)
 for each work/row i:
@@ -32,7 +34,7 @@ for each work/row i:
 }
 \examples{
 \dontrun{
-flat_snow <- to_disk(oa_snowball(
+flat_snow <- snowball2df(oa_snowball(
   identifier = "W1516819724",
   verbose = TRUE
 ))
