+++ man/oa_snowball.Rd
@@ -30,8 +30,8 @@ for example, https://openalex.org/W2755950973}
 \item{endpoint}{Character. URL of the OpenAlex Endpoint API server.
 Defaults to endpoint = "https://api.openalex.org/".}
 
-\item{verbose}{Logical. If TRUE, information about the querying process will
-be plotted on screen. Default to \code{verbose = FALSE}.}
+\item{verbose}{Logical. If TRUE, print information on querying process.
+Default to \code{verbose = FALSE}.}
 
 \item{citing_filter}{filters used in the search of works citing the input works.}
 
