### 2022-10-10 dc18416
to_disk is now snowball2df to be consistent with the oa2df family of functions
