+++ NAMESPACE
@@ -10,5 +10,5 @@ export(oa_request)
 export(oa_snowball)
 export(show_authors)
 export(show_works)
-export(to_disk)
+export(snowball2df)
 importFrom(stats,setNames)
