+++ R/oa_fetch.R
@@ -56,9 +56,9 @@ oa_entities <- function() {
 #'   verbose = TRUE
 #' )
 #' }
-oa_fetch <- function(...,
-                     identifier = NULL, ## identifier of a work, author, venue, etc.
+oa_fetch <- function(identifier = NULL, ## identifier of a work, author, venue, etc.
                      entity = if (is.null(identifier)) NULL else id_type(identifier[[1]]),
+                     ...,
                      search = NULL,
                      sort = NULL,
                      group_by = NULL,
