+++ README.md
@@ -75,7 +75,7 @@ theme_set(theme_classic())
 ## Examples
 
 There are different
-[filters](https://docs.openalex.org/api/get-lists-of-entities/filter-entity-lists)/arguments
+[filters](https://massimoaria.github.io/openalexR/articles/Filters)/arguments
 you can use in `oa_fetch`, depending on which
 [entity](https://docs.openalex.org/about-the-data) you’re interested in:
 works, authors, venues, institutions, or concepts. We show a few
@@ -87,7 +87,7 @@ examples below.
 (known DOIs).
 
 Use `doi` as a [works
-filter](https://docs.openalex.org/api/get-lists-of-entities/filter-entity-lists#works-filters)
+filter](https://massimoaria.github.io/openalexR/articles/Filters.html#works)
 (either the canonical form with <https://doi.org/> or without):
 
 ``` r
@@ -125,14 +125,14 @@ oa_fetch(
 #> About to get a total of 2 pages of results with a total of 211 records.
 ```
 
-| id          | display_name                                                                                                                              | first_author          | last_author         | so                              | url                                             | is_oa | top_concepts                                                |
-|:------------|:------------------------------------------------------------------------------------------------------------------------------------------|:----------------------|:--------------------|:--------------------------------|:------------------------------------------------|:------|:------------------------------------------------------------|
-| W2755950973 | bibliometrix : An R-tool for comprehensive science mapping analysis                                                                       | Massimo Aria          | Corrado Cuccurullo  | Journal of Informetrics         | <https://doi.org/10.1016/j.joi.2017.08.007>     | FALSE | Data science, Information retrieval                         |
-| W2955219525 | Scaling tree-based automated machine learning to biomedical big data with a feature set selector                                          | Trang T. Le           | Jason H. Moore      | Bioinformatics                  | <https://doi.org/10.1093/bioinformatics/btz470> | TRUE  | Pipeline (software), Scalability, Feature (linguistics)     |
-| W1979874437 | Analysis of powered two-wheeler crashes in Italy by classification trees and rules discovery                                              | Alfonso Montella      | Filomena Mauriello  | Accident Analysis & Prevention  | <https://doi.org/10.1016/j.aap.2011.04.025>     | FALSE | Crash, Identification (biology), Decision tree              |
-| W2952824318 | A Nonlinear Simulation Framework Supports Adjusting for Age When Analyzing BrainAGE                                                       | Trang T. Le           | Tulsa Investigators | Frontiers in Aging Neuroscience | <https://doi.org/10.3389/fnagi.2018.00317>      | TRUE  | Correlation, Mood, Contrast (vision)                        |
-| W2408216567 | Foundations and trends in performance management. A twenty-five years bibliometric analysis in business and public administration domains | Corrado Cuccurullo    | Fabrizia Sarto      | Scientometrics                  | <https://doi.org/10.1007/s11192-016-1948-8>     | FALSE | Administration (probate law), Bibliometrics, Trend analysis |
-| W2281330131 | Coopetition and sustainable competitive advantage. The case of tourist destinations                                                       | Valentina Della Corte | Massimo Aria        | Tourism Management              | <https://doi.org/10.1016/j.tourman.2015.12.009> | FALSE | Competitive advantage, Tourism                              |
+| id          | display_name                                                                                                                              | first_author          | last_author         | so                              | url                                             | is_oa | top_concepts                                                   |
[…]
