+++ README.Rmd
@@ -64,7 +64,17 @@ Bonus point if you put this in your `.Rprofile` with `file.edit("~/.Rprofile")`.
 library(openalexR)
 library(dplyr)
 library(ggplot2)
-theme_set(theme_classic())
+```
+
+```{r set-theme, include=FALSE}
+theme_set(
+  theme_classic() +
+    theme(
+      plot.background = element_rect(fill = "transparent", colour = NA),
+      panel.background = element_rect(fill = "transparent", colour = NA)
+    )
+)
+knitr::opts_chunk$set(dev.args = list(bg = "transparent"))
 ```
 
 ## Examples
@@ -231,18 +241,18 @@ italy_insts %>%
 And what do they publish on?
 
 ```{r concept-cloud, fig.height=5, fig.width=7}
-concept_cloud <- italy_insts %>% 
-  select(inst_id = id, x_concepts) %>% 
-  tidyr::unnest(x_concepts) %>% 
-  filter(level == 1) %>% 
-  select(display_name, score) %>% 
-  group_by(display_name) %>% 
+concept_cloud <- italy_insts %>%
+  select(inst_id = id, x_concepts) %>%
+  tidyr::unnest(x_concepts) %>%
+  filter(level == 1) %>%
+  select(display_name, score) %>%
+  group_by(display_name) %>%
   summarise(score = sum(score))
 
 pal <- c("black", scales::brewer_pal(palette = "Set1")(5))
 set.seed(1)
 wordcloud::wordcloud(
-  concept_cloud$display_name, 
+  concept_cloud$display_name,
   concept_cloud$score,
   scale = c(2, .4),
   colors = pal
@@ -285,6 +295,8 @@ jours %>%
   coord_polar(clip = "off") +
   theme_bw() +
   theme(
+    plot.background = element_rect(fill = "transparent", colour = NA),
+    panel.background = element_rect(fill = "transparent", colour = NA),
     panel.grid = element_blank(),
     panel.border = element_blank(),
     axis.text = element_blank(),
