+++ pkgdown.yml
@@ -4,8 +4,9 @@ pkgdown_sha: ~
 articles:
   A_Brief_Introduction_to_openalexR: A_Brief_Introduction_to_openalexR.html
   About-the-tibble-output: About-the-tibble-output.html
+  Filters: Filters.html
   institution: institution.html
-last_built: 2022-10-07T19:20Z
+last_built: 2022-10-11T12:18Z
 urls:
   reference: https://massimoaria.github.io/openalexR/reference
   article: https://massimoaria.github.io/openalexR/articles
