### 2022-10-17 c3730d3
small cleanup
