+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   About-the-tibble-output: About-the-tibble-output.html
   Filters: Filters.html
   institution: institution.html
-last_built: 2022-10-17T16:29Z
+last_built: 2022-10-19T13:01Z
 urls:
   reference: https://massimoaria.github.io/openalexR/reference
   article: https://massimoaria.github.io/openalexR/articles
