### 2022-10-27 #557 @adamhsparks
Hi @trangdata, could you give us a more detailed comparison of the packages' overlap?

Here's an example of a previous inquiry for another package that had some overlap. https://github.com/ropensci/software-review/issues/199#issuecomment-375358362
