### 2022-10-28 #557 @trangdata
Thank you @adamhsparks for clarifying! 🌻

I'm showing a detailed table of comparison below. In short, our package shares some similarity with KTH-Library/openalex package, but we offer additional functionalities such as [snowball search](https://www.wohlin.eu/ease14.pdf). We also expose functions that are crucial for many use cases such as filtering for a particular set of works/authors/etc. and generally a more "canonical" and simpler way to construct the query (see example). Our package has been well maintained and well tested. We also have a lot fewer dependencies.

Let me know if there is anything you would like me to elaborate on. 🌈

## Table of comparison
| Type     |        Property                                            | massimoaria/openalexR                  |        KTH-Library/openalex                                     | ekmaloney/openalexR                                                  |
|----------|------------------------------------------------------------|----------------------------------------|-----------------------------------------------------------------|----------------------------------------------------------------------|
| Metadata | last commit                                                | 11 days ago                            | 8 months ago                                                    | 9 months ago                                                         |
| Metadata | # vignettes                                                | 4                                      | 0                                                               |                                                                      |
| Metadata | # readme examples                                          | 10                                     | 1                                                               |                                                                      |
| Metadata | test coverage                                              | 89%                                    | N/A                                                             | N/A                                                                  |
| Metadata | # commits                                                  | 198                                    | 7                                                               | 20                                                                   |
| Metadata | # pull requests                                            | 17                                     | 0                                                               | 0                                                                    |
| Metadata | # stars                                                    | 26                                     | 8                                                               | 7                                                                    |
| Metadata | license                                                    | MIT                                    | MIT                                                             | MIT                                                                  |
| Metadata | Imports                                                    | curl, httr, jsonlite, progress, tibble | httr, magrittr, utils, dplyr, purrr, progress, jsonlite, tibble | magrittr, dplyr, httr, jsonlite, purrr, tidyr, stringr, tibble       |
| Function | request data from query                                    | oa_request                             |        openalex_crawl                                           |        openalex_api                                                  |
| Function | enter polite pool                                          | openalexR.mailto                       | openalex_polite                                                 |                                                                      |
| Function | generate a valid query                                     | oa_query                               | openalex_query                                                  | construct_links                                                      |
| Function | converts the JSON object in classical bibliographic tibble | oa2df                                  | openalex_flatten_long                                           | clean_author_info, clean_venue_info, clean_works_info                |
| Function | composes functions to ease data fetching                   | oa_fetch                               | N/A                                                             | find_work, find_author, find_institution, find_concept, find_venue   |
| Function | display openalex attribution                               |                                        | openalex_attribution                                            |                                                                      |
| Function | snowballing                                                | oa_snowball                            |                                                                 |                                                                      |
| Function | convert works to bibliometrix object                       | oa2bibliometrix                        |                                                                 |                                                                      |
| Function | get random entity                                          | oa_random                              |                                                                 |                                                                      |
| Function | simplify works result                                      | show_works                             |                                                                 |                                                                      |
| Function | simplify authors result                                    | show_authors                           |                                                                 |                                                                      |
| Function | available entities in the OpenAlex database                | oa_entities                            |                                                                 |                                                                      |
| Function | get papers for a particular author                         | oa_fetch (see example)                 | openalex_flatten_long (see example)                             | get_authors_papers                                                   |
| Function | get coauthors for a particular   author                    |                                        |                                                                 | get_coauthors                                                        |
| Function | other functions, unknown use case                          |                                        |                                                                 | get_links_for_each_page, get_number_of_pages, get_all_data_for_query |
| Data     | list of countries and their alpha-2 and aplha-3 codes      | countrycode                            |                                                                 |                                                                      |
| Data     |        0-level   concepts and corresponding abbreviations  | concept_abbrev                         |                                                                 |                                                                      |

## Example

Getting a dataframe of papers published in 2022 by a specific author.

### Our way
``` r
oa_fetch(
  "works",
  authors.orcid = "0000-0002-8517-9411",
  publication_year = 2022
)
```

### KTH-Library/openalex's way:
``` r
openalex_flatten_long(openalex_crawl(
  "works",
  query = openalex:::openalex_query(
    filter = "authors.orcid:0000-0002-8517-9411,publication_year:2022"
  )
))
```

### ekmaloney/openalexR's way:
Not available. The package doesn't allow for filtering by publication_year.
