### 2022-11-05 #557 @trangdata
Hi @adamhsparks — just wanted to float this up on your list. I'm excited to get the package to ropensci and start the review process. Let me know if there is anything else that you need from me, or if there is anything from the [presubmission guide](https://devguide.ropensci.org/authors-guide.html#planning-a-submission-or-a-pre-submission-enquiry) that I should highlight. Thank you again for your time! 🪴
